requirements.txt
matplotlib==3.3.4
numpy==1.21.4
pandas==1.2.4
Pillow==8.4.0
streamlit==1.2.0
streamlit_drawable_canvas==0.8.0
tensorflow-cpu==2.7.0
